Add Word Wrap Bug
Sets the style of wrapping used if the text area is wrapping lines.
If set to true the lines will be wrapped at word boundaries (whitespace)
if they are too long to fit within the allocated width.
If set to false, the lines will be wrapped at character boundaries. By
default this property is false.